src/requirements.txt
langchain>=0.1.0
langchain-google-vertexai
langgraph
deepagents
pydantic
google-cloud-secret-manager
google-auth
pdfplumber
PyMuPDF
boto3
